+++ lib/twenv/commands/archive_timeline.rb
@@ -28,7 +28,10 @@ class TWEnv::ArchiveTimeline < TWEnv::Command
   rescue Interrupt
     line.end_line
   ensure
-    puts "Archive saved to #{@path}"
+    locals = pry_instance.config.extra_sticky_locals
+    locals.merge!(archived_timeline: JSON.parse(File.read(@path)))
+    line.print "Archive saved to #{@path}"
+    line.end_line.print("Archive assigned to local variable `archived_timeline`").end_line
   end
 
   def options(slop)